'never push buggy code' oops

+++ src/breakout/Main.java
@@ -1,29 +1,24 @@
 package breakout;
 
-import javafx.scene.Node;
 import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.control.Label;
+import javafx.scene.layout.VBox;
 import javafx.scene.paint.Color;
 import javafx.scene.paint.Paint;
-import javafx.scene.shape.Rectangle;
 import javafx.animation.KeyFrame;
 import javafx.animation.Timeline;
 import javafx.application.Application;
 import javafx.scene.Group;
 import javafx.scene.input.KeyCode;
 import javafx.scene.shape.Shape;
+import javafx.scene.text.Font;
+import javafx.scene.text.Text;
 import javafx.stage.Stage;
 import javafx.util.Duration;
-import javafx.scene.shape.Circle;
 
 import java.io.File;
-import java.io.FileNotFoundException;
-import java.util.Scanner;
 
-//import java.lang.invoke.DelegatingMethodHandle$Holder;
-
-/**
- * Feel free to completely change this code or delete it entirely. 
- */
 public class Main extends Application{
     /**
      * Start of the program.
@@ -35,17 +30,17 @@ public class Main extends Application{
     public static final double SECOND_DELAY = 1.0 / FRAMES_PER_SECOND;
     public static final double TEN_SECONDS = SECOND_DELAY*10;
     public static final Paint BACKGROUND = Color.AZURE;
-    //public static final Paint HIGHLIGHT = Color.LAVENDERBLUSH;
     public static final int GAP = 10;
     public static final int BRICK_WIDTH = 20;
     public static final int BRICK_HEIGHT = 20;
-    //public static final Paint BALL_COLOR = Color.GREY;
     public static final int BALL_RADIUS = 10;
     public static final int BASIC_BALL_SPEED = 100;
     public static final int DOUBLE_IN_SIZE = 2;
     public static final int STARTING_LIVES = 3;
 
     private Scene myScene;
+    private SplashScreen mySplash;
+    private Scene holder;
     private Paddle myPaddle;
     private Ball myBall;
     private Ball bonusBall;
@@ -53,14 +48,15 @@ public class Main extends Application{
     private int myLives;
     //private int myLevelNumber;
     private int myScore;
-    private Level myLevel;
-    //private Brick myBrick;
+    private Level myLevel1;
+    private Level myLevel2;
+    private Level myLevel3;
+    private Laser myLaser;
+    private boolean laserIsOnBoard = false;
     private boolean paddleIsNormal;
     private Group root;
     private boolean paddleCanMove;
-    private boolean brickHasBeenDeleted;
     private boolean bonusBallExists = false;
-
 
 // ====================================================================================================================
 // beginning the game
@@ -69,93 +65,54 @@ public class Main extends Application{
     @Override
     public void start (Stage stage) {
         // attach scene to the stage and display it
-        myScene = setupGame(SIZE, SIZE, BACKGROUND);
-        stage.setScene(myScene);
+        mySplash = new SplashScreen();
+        stage.setScene(mySplash.getNode());
         stage.setTitle(TITLE);
+        myLives = 3;
+        myScore = 0;
+        //mySplash.getNode().setOnKeyPressed(e -> handleKeyInput(e.getCode()));
+        //if()
+        myScene = setupGame(SIZE, SIZE, BACKGROUND, "level1");
+        stage.setScene(myScene);
+
         stage.show();
+
+        Level myLevel = setUpLevel("level1");
+        addNewLevelToRoot(myLevel);
+
         // attach "game loop" to timeline to play it (basically just calling step() method repeatedly forever)
-        KeyFrame frame = new KeyFrame(Duration.millis(MILLISECOND_DELAY), e -> step(SECOND_DELAY));
+        KeyFrame frame = new KeyFrame(Duration.millis(MILLISECOND_DELAY), e -> step(SECOND_DELAY, myLevel));
         Timeline animation = new Timeline();
         animation.setCycleCount(Timeline.INDEFINITE);
         animation.getKeyFrames().add(frame);
         animation.play();
     }
-    private Scene setupGame (int width, int height, Paint background) {
+    private Scene setupGame (int width, int height, Paint background, String levelName) {
         // create one top level collection to organize the things in the scene
         /*Group*/ root = new Group();
-        myLives = 3; //TODO set up this in each level
-        myScore = 0;
-        //myLevelNumber = 1;
+        //myLives = 3; //TODO set up this in each level
+        //myScore = 0;
 
-        myBall = new Ball();
-        //myBall.setBallInformation(BALL_COLOR, SIZE, BALL_RADIUS, GAP, PADDLE_HEIGHT, BASIC_BALL_SPEED);
-        //myBall.newBall();
         myBall = new Ball();
         bonusBallExists = false;
-
-/*
-        myBrick = new Brick();
-        myBrick.setBrickInformation(BRICK_WIDTH, BRICK_HEIGHT, SIZE, BACKGROUND);
-        myBrick.makeSingleBrick(2,20,0);
-        brickHasBeenDeleted = (myBrick.getHitsRemaining()==0); //true if hits = 0, false otherwise
-*/
-        File filename = new File("/Users/danamulligan/Archives/workspace308/game_dmm107/doc/level1");//CHANGE THIS LATER SO YOU DONT DIE WHEN SUBMITTING
-        /*try{
-            Scanner sc = new Scanner(filename); //red line bc you're not checking that filename is a valid file
-            for(int row=0; row < SIZE/BRICK_HEIGHT; row++){
-                for(int col=0; col< SIZE/BRICK_WIDTH; col++){
-                    if(sc.hasNext()){
-                        hitsNeeded[row][col] = sc.nextInt();
-                        locationInfoX[row][col] = col*BRICK_WIDTH;
-                        locationInfoY[row][col] = row*BRICK_HEIGHT;
-                        //System.out.println("Hits Needed: " + hitsNeeded[row][col] + " X: "+locationInfoX[row][col]+" Y: "+locationInfoY[row][col]);
-                        totalNumberOfBricks++;
-                    }
-                }
-            }
-        }catch (FileNotFoundException e){
-            e.printStackTrace();
-        }
-        */
-        myLevel = new Level();
-        myLevel.setLevelInformation(filename, SIZE, BRICK_WIDTH, BRICK_HEIGHT);
-        myLevel.setUpLevel();
-
+        root.getChildren().add(myBall.getNode());
 
         myPaddle = new Paddle();
         paddleCanMove = false;
         paddleIsNormal = true;
-        //myPaddle.setPaddleInformation(SIZE, PADDLE_WIDTH, PADDLE_HEIGHT, GAP, PADDLE_COLOR, PADDLE_SPEED);
-        myPaddle = new Paddle();
-
-        /*myLevel = new Level();
-        //TODO change the path below
-
-        myLevel.setLevelInformation(levelNumber, SIZE, BRICK_WIDTH, BRICK_HEIGHT);
-        myLevel.setUpLevel();
-        addNewLevelToRoot(myLevel);
-
-
-         */
-        //setLevelInformation(levelNumber);
-       // printBoard();
-
-        /*
-        System.out.println("trying to add myBricks[0][1] to root; it takes "+myBricks[0][1].getHitsRemaining()+" to break");
-        Rectangle rect = myBricks[0][1].getNode();
-        root.getChildren().add(rect);
-
-         */
-        //bonusBallExists = true;
-        //bonusBall = new Ball();
-        //bonusBall.setBallInformation(Color.PURPLE, SIZE, BALL_RADIUS, GAP, PADDLE_HEIGHT, BASIC_BALL_SPEED);
-        //bonusBall.newBonusBall(BALL_RADIUS, myBall.getNode().getCenterX(), myBall.getNode().getCenterY(), myBall.getBallXSpeed(), myBall.getBallYSpeed());
-        //root.getChildren().add(bonusBall.getNode());
-
-        // order added to the group is the order in which they are drawn
-        root.getChildren().add(myBall.getNode());
         addPaddleToRoot();
-        addNewLevelToRoot(myLevel);
+
+        /*myLevel1 = setUpLevel("level1");
+        addNewLevelToRoot(myLevel1);
+        myLevel2 = setUpLevel("level2");
+        addNewLevelToRoot(myLevel2);
+        myLevel3 = setUpLevel("level3");
+        addNewLevelToRoot(myLevel3);*/
+        /*
+        myLevel = setUpLevel(levelName);
+        addNewLevelToRoot(myLevel);*/
+
+        //root.getChildren().add(bonusBall.getNode());
 
         // create a place to see the shapes
         Scene scene = new Scene(root, width, height, background);
@@ -164,16 +121,24 @@ public class Main extends Application{
         //scene.setOnMouseClicked(e -> handleMouseInput(e.getX(), e.getY()));
         return scene;
     }
+    private Level setUpLevel(String levelName){
+        File filename = new File(this.getClass().getClassLoader().getResource(levelName).getFile());
+        return new Level(filename);
+    }
 // ====================================================================================================================
 // playing the game
 // ====================================================================================================================
     // Change properties of shapes in small ways to animate them over time
     // Note, there are more sophisticated ways to animate shapes, but these simple ways work fine to start
-    private void step (double elapsedTime) {
+    private void step (double elapsedTime, Level myLevel) {
         // move ball
-
         //bounce ball on paddle
-        hitBrick();
+        hitBricks(myBall.getNode(), "ball", myLevel);
+        if(laserIsOnBoard) {
+            hitBricks(myLaser.getLeftLaserBeamNode(), "laser left", myLevel);
+            hitBricks(myLaser.getRightLaserBeamNode(), "laser right", myLevel);
+            myLaser.moveLasers(elapsedTime);
+        }
         moveBall(myBall, elapsedTime);
         bounceBallOffPaddle(myBall);
         if(bonusBallExists){
@@ -191,45 +156,62 @@ public class Main extends Application{
     private void bounceBallsOffWalls(){
         if(myBall.ballCenterX()+BALL_RADIUS>=SIZE || myBall.ballCenterX()<=0) {
             myBall.bounceSideWall();
+            //myBall.resetHitCapabilities();
         }
         if(myBall.ballCenterY()<=0) {
             myBall.bounceY();
+            myBall.resetHitCapabilities();
         }
         if(myBall.ballCenterY()+BALL_RADIUS>=SIZE) {
             myLives--;
             resetGame();
-        }
+        }/*
         if(bonusBallExists && (bonusBall.ballCenterX()+BALL_RADIUS>=SIZE || bonusBall.ballCenterX()<=0)) {
             bonusBall.bounceSideWall();
+            //myBall.resetHitCapabilities();;
         }
         if(bonusBallExists && bonusBall.ballCenterY()<=0) {
             bonusBall.bounceY();
-        }
+            myBall.resetHitCapabilities();
+        }*/
         if(myLives == 0){
             endGame();
         }
     }
-    private void hitBrick(){
+    private void hitBricks(Shape breaker, String label, Level myLevel){
         //TODO change this to send it down into level class
         Brick myBrick;
-        for(int row = 0; row<SIZE/Brick.brickHeight; row++) {
-            for (int col = 0; col < SIZE/Brick.brickWidth; col++) {
-                myBrick = myLevel.myBricks[row][col];
-                if (!myBrick.isBrickDestroyed()){
-                    Shape intersectionBrick = Shape.intersect(myBall.getNode(), myBrick.getNode());
-                    if (intersectionBrick.getBoundsInLocal().getWidth() != -1 && myBrick.getHitsRemaining()!=0) {
-                        myBall.bounceY();
-                        myBrick.brickIsHit();
-                        myScore += 100;
-                        System.out.println("HIT!");
-                        System.out.println(myBrick.getHitsRemaining() + " hits remain!");
-                        if (myBrick.getHitsRemaining() == 0) {
-                            myBrick.deleteBrick();
-                            root.getChildren().remove(myBrick.getNode());
-                            System.out.println("this is not infinite loop!");
-                        }
-                    }
 
+        for(int row = 0; row<SIZE/Brick.BRICK_HEIGHT; row++) {
+            for (int col = 0; col < SIZE/Brick.BRICK_WIDTH; col++) {
+                myBrick = myLevel.myBricks[row][col];
+                boolean breakerHitStatus = getBreakerHitStatus(label);
+                if (!myBrick.isBrickDestroyed() && breakerHitStatus){//&&!myBrick.getHitStatus()){
+                    Shape intersectionBrick = Shape.intersect(breaker, myBrick.getNode());
+                    if(intersectionBrick.getBoundsInLocal().getWidth() != -1 /*&& myBrick.getHitsRemaining()!=0 && !myBrick.getHitStatus()*/) {
+                        if(label.equals("ball")){
+                            myBall.bounceY();
+                            //myBall.bounceSideBrick();
+                            myBall.ballHitABrick();
+                        }
+                        myBrick.brickIsHit();
+                        if(label.equals("laser left")) {
+                            myLaser.leftLaserHitABrick();
+                            //root.getChildren().remove(myLaser.getLeftLaserBeamNode());
+                            myLaser.destroyLaserBeamLeft();
+                        }
+                        if(label.equals("laser right")){
+                            myLaser.rightLaserHitABrick();
+                            //root.getChildren().remove(myLaser.getRightLaserBeamNode());
+                            myLaser.destroyLaserBeamRight();
+                        }
+                        //myBrick.resetHitStatusToFalse();
+                        myScore += 100;
+                    }
+                    if (myBrick.getHitsRemaining() == 0) {
+                        myBrick.deleteBrick();
+                        root.getChildren().remove(myBrick.getNode());
+                    }
                 }
             }
         }
@@ -240,15 +222,19 @@ public class Main extends Application{
         Shape intersectionRight = Shape.intersect(ballToBounce.getNode(), myPaddle.getNodeRight());
         if (intersectionLeft.getBoundsInLocal().getWidth() != -1) {
             ballToBounce.bouncePaddleLeft();
+            //myBall.resetHitCapabilities();
         }
         if (intersectionMiddle.getBoundsInLocal().getWidth() != -1) {
             ballToBounce.bouncePaddleMiddle();
+            //myBall.resetHitCapabilities();
         }
         if (intersectionRight.getBoundsInLocal().getWidth() != -1) {
             ballToBounce.bouncePaddleRight();
+            //myBall.resetHitCapabilities();
         }
         if((intersectionLeft.getBoundsInLocal().getWidth()!=-1 || intersectionRight.getBoundsInLocal().getWidth()!=-1) && intersectionMiddle.getBoundsInLocal().getWidth()!=-1){
             ballToBounce.bounceY();
+           // myBall.resetHitCapabilities();
         }
     }
 // ====================================================================================================================
@@ -277,7 +263,7 @@ public class Main extends Application{
         }
         else if (code == KeyCode.S && !paddleIsNormal){ //shrink
             removePaddleFromRoot();
-            myPaddle = new Paddle(DOUBLE_IN_SIZE, "shrink");
+            myPaddle = new Paddle();//DOUBLE_IN_SIZE, "shrink");
             paddleIsNormal = true;
             addPaddleToRoot();
         }
@@ -287,19 +273,12 @@ public class Main extends Application{
         else if (code == KeyCode.T){ //slow like a turtle
             myBall.changeBallSpeed(DOUBLE_IN_SIZE, "slow");
         }
-        //else if (code == KeyCode.V){
-        //    myBall.growBall();
-        //}/*
-        //else if (code == KeyCode.B){
-          //  bonusBallExists = true;
-           // bonusBall = new Ball();
-           // bonusBall.setBallInformation(Color.PURPLE, SIZE, GAP, PADDLE_HEIGHT, BASIC_BALL_SPEED);
-          //  bonusBall.newBonusBall(BALL_RADIUS, myBall.getNode().getCenterX(), myBall.getNode().getCenterY(), myBall.getBallXSpeed(), myBall.getBallYSpeed());
-           // root.getChildren().add(bonusBall.getNode());
-           // System.out.println("there should be a bonus ball!");
-        //}*/
+        else if (code == KeyCode.Z){
+            myLaser = new Laser(myPaddle.getLeftX(),myPaddle.getActualPaddleWidth());
+            laserIsOnBoard = true;
+            root.getChildren().addAll(myLaser.getLeftLaserBeamNode(), myLaser.getRightLaserBeamNode());
+        }
         else if (code == KeyCode.L){ //add a life
-            //TODO
             myLives++;
             System.out.println("Plus one life! You now have "+myLives+" lives");
         }
@@ -350,11 +329,22 @@ public class Main extends Application{
         myPaddle = new Paddle();
         addPaddleToRoot();
     }
+    public boolean getBreakerHitStatus(String label){
+        if(label.equals("ball")){
+            return myBall.getHitStatus();
+        }
+        else if(label.equals("laser left")){
+            return myLaser.getLeftLaserHitStatus();
+        }
+        else if(label.equals("laser right")){
+            return myLaser.getRightLaserHitStatus();
+        }
+        return false;
+    }
     public void addNewLevelToRoot(Level levelToBeAdded){
         for(int row=0; row<SIZE/BRICK_HEIGHT; row++){
             for(int col=0; col<SIZE/BRICK_WIDTH; col++){
                 root.getChildren().add(levelToBeAdded.getBrickNode(row,col));
-                //System.out.println("added brick at x: "+col+ " and y: "+row+" to root");
             }
         }
     }
@@ -362,17 +352,13 @@ public class Main extends Application{
      * reason why I have this method is because it's "one" object, and I want to read it as such
      */
     public void removePaddleFromRoot(){
-        root.getChildren().remove(myPaddle.getNodeLeft());
-        root.getChildren().remove(myPaddle.getNodeMiddle());
-        root.getChildren().remove(myPaddle.getNodeRight());
+        root.getChildren().removeAll(myPaddle.getNodeLeft(), myPaddle.getNodeMiddle(), myPaddle.getNodeRight());
     }
     /**
      * reason why I have this method is because it's "one" object, and I want to read it as such
      */
     public void addPaddleToRoot(){
-        root.getChildren().add(myPaddle.getNodeLeft());
-        root.getChildren().add(myPaddle.getNodeMiddle());
-        root.getChildren().add(myPaddle.getNodeRight());
+        root.getChildren().addAll(myPaddle.getNodeLeft(), myPaddle.getNodeMiddle(), myPaddle.getNodeRight());
     }
     public static void main (String[] args) {
         launch(args);
